refactor: update StakeCard componenet
